refactor : 안쓰는 코드 삭제 및 Qualifier 수정
